Remove Producer report query in existing Producer repo and created new stand-alone repo to make it consistent with existing reports.
