+++ src/EA.Iws.DataAccess/EA.Iws.DataAccess.csproj
@@ -288,6 +288,7 @@
     <Compile Include="Repositories\Reports\FinanceReportRepository.cs" />
     <Compile Include="Repositories\Reports\ImportNotificationsRepository.cs" />
     <Compile Include="Repositories\Reports\ImportStatsRepository.cs" />
+    <Compile Include="Repositories\Reports\ProducerRepository.cs" />
     <Compile Include="Repositories\Reports\ShipmentsRepository.cs" />
     <Compile Include="Repositories\Reports\FreedomOfInformationRepository.cs" />
     <Compile Include="Repositories\NumberOfShipmentsHistoryRepository.cs" />
